厳格なコンパイラ警告オプションを追加し、警告ゼロポリシーを強制
変更内容:
- SAPPP_WERROR オプションを追加（警告をエラーとして扱う）
- GCC/Clang/MSVC 向けに包括的な警告オプションを追加:
  -Wall -Wextra -Wpedantic -Wshadow -Wnon-virtual-dtor
  -Wold-style-cast -Wcast-align -Wunused -Woverloaded-virtual
  -Wconversion -Wsign-conversion -Wnull-dereference
  -Wdouble-promotion -Wformat=2 -Wimplicit-fallthrough
  (GCC専用: -Wduplicated-cond -Wduplicated-branches
   -Wlogical-op -Wuseless-cast)
- sappp_target_strict_warnings() 関数でプロジェクトコードにのみ適用
- サードパーティライブラリは SYSTEM include で警告を抑制
- sha256.cpp: ループインデックスを int から size_t に修正
- main.cpp: jobs 変数に [[maybe_unused]] を追加
- AGENTS.md: 警告ゼロポリシーを必須事項に追加
- AGENTS.md: コミットメッセージは日本語で書くルールを追加

+++ libs/common/sha256.cpp
@@ -113,7 +113,7 @@ public:
 
         // Output (big-endian)
         std::array<uint8_t, 32> hash;
-        for (int i = 0; i < 8; ++i) {
+        for (size_t i = 0; i < 8; ++i) {
             hash[i * 4 + 0] = static_cast<uint8_t>(state_[i] >> 24);
             hash[i * 4 + 1] = static_cast<uint8_t>(state_[i] >> 16);
             hash[i * 4 + 2] = static_cast<uint8_t>(state_[i] >> 8);
@@ -127,13 +127,13 @@ private:
         std::array<uint32_t, 64> w;
 
         // Prepare message schedule
-        for (int i = 0; i < 16; ++i) {
+        for (size_t i = 0; i < 16; ++i) {
             w[i] = (static_cast<uint32_t>(buffer_[i * 4 + 0]) << 24) |
                    (static_cast<uint32_t>(buffer_[i * 4 + 1]) << 16) |
                    (static_cast<uint32_t>(buffer_[i * 4 + 2]) << 8) |
                    (static_cast<uint32_t>(buffer_[i * 4 + 3]));
         }
-        for (int i = 16; i < 64; ++i) {
+        for (size_t i = 16; i < 64; ++i) {
             w[i] = gamma1(w[i - 2]) + w[i - 7] + gamma0(w[i - 15]) + w[i - 16];
         }
 
@@ -142,7 +142,7 @@ private:
         uint32_t e = state_[4], f = state_[5], g = state_[6], h = state_[7];
 
         // Compression
-        for (int i = 0; i < 64; ++i) {
+        for (size_t i = 0; i < 64; ++i) {
             uint32_t t1 = h + sigma1(e) + ch(e, f, g) + K[i] + w[i];
             uint32_t t2 = sigma0(a) + maj(a, b, c);
             h = g; g = f; f = e; e = d + t1;
